% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimations.R
\name{f.bid_function_fast}
\alias{f.bid_function_fast}
\title{Suite of functions to estimate private-value auction models}
\usage{
f.bid_function_fast(cost, num_bids, mu, alpha, gamma_1p1oa)
}
\arguments{
\item{cost}{XXnumber of observations to draw}

\item{num_bids}{XXnon-negative alpha parameter of the beta distribution}

\item{mu}{XXnon-negative beta parameter of the beta distribution}

\item{alpha}{XXnon-negative beta parameter of the beta distribution}

\item{gamma_1p1oa}{XXnon-negative beta parameter of the beta distribution}
}
\description{
Suite of functions to estimate private-value auction models
}
\details{
The Beta distribution with parameters \eqn{a} and \eqn{b} has
density:

\deqn{
    \Gamma(a+b)/(\Gamma(a)\Gamma(b))x^(a-1)(1-x)^(b-1)
}

for \eqn{a > 0}, \eqn{b > 0} and \eqn{0 \le x \le 1}.
}
\examples{
# Draw from beta distribution with parameters a = 1 and b = 3
beta_plot(a = 1, b = 3)

}
\seealso{
\code{\link{rbeta}}, \code{\link{geom_density}}
}
