% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimations.R
\name{print.auctionmodel}
\alias{print.auctionmodel}
\title{Print an auction model.}
\usage{
\method{print}{auctionmodel}(x, digits = 6, ...)
}
\arguments{
\item{x}{An object of class auctionmodel.}

\item{digits}{Number of digits to display.}

\item{...}{Additional arguments passed to other methods.}
}
\value{
x, invisibly.
}
\description{
Print an auction model.
}
