% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimations.R
\name{auction_model}
\alias{auction_model}
\title{Estimates a first-price auction model}
\usage{
auction_model(
  dat = NULL,
  init_param = NULL,
  num_cores = 1,
  method = "BFGS",
  control = list()
)
}
\arguments{
\item{dat}{data.frame containing input columns in the following order: the winning bids, number of bids, and \code{X} variables that represent observed heterogeneity.}

\item{num_cores}{The number of cores for running the model in parallel.}

\item{method}{Optimization method to be used in optim() (see ?optim for details).}

\item{control}{A list of control parameters to be passed to optim() (see ?optim for details).}

\item{init_params}{Vector of initial values for mu, alpha, sigma, and beta vector.}
}
\description{
Estimates a first-price auction model
}
\details{
This function estimates a first-price auction model with conditional independent private values.
The model allows for unobserved heterogeneity that is common to all bidders in addition to observable
heterogeneity. The winning bid (Y) takes the form

Y = B * U * h(X)

where B Is the proportional winning bid, U is the unobserved heterogeneity, and h(X) controls for
observed heterogeneity. The model is log-linear so that
log(Y) = log(B) + log(U) + log(h(X)) and log(h(X)) = beta1 * X1 + beta2 * X2 + ...

The (conditionally) independent private costs are drawn from a Weibull distribution
with parameters mu and alpha. The CDF of this distribution is given by

F(c) = 1 - exp(- (c * 1/mu * Gamma(1 + 1/alpha))^(alpha))

The unobserved heterogeneity U is sampled from log-Normal distribution with mean 1 and a free parameter sigma representing its standard deviation.

\code{init_params}, the initial guess for convergence, must be supplied.

This funtion utilizes the \code{Rsnow} framework within the \code{Rparallel} package. If \code{numcores} is not specified, this will be run using only
one CPU/core. One can use \code{parallel::detectCores()} to determine how many are available on your system, but you are not advised
to use all at once, as this may make your system unresponsive. Please see \code{Rparallel} and \code{Rsnow} for more details.

Note that the supplied data can not have missing values.
}
\examples{

set.seed(100)
dat <- auction_generate_data(obs = 100, mu = 10, alpha = 2, sigma = 0.2,
                             beta = c(-1,1), new_x_mean= c(-1,1), new_x_sd = c(0.5,0.8))
auction_model(dat,
              init_param =  c(8, 2, .5, .4, .6),
              num_cores = 1,
              method = "BFGS",
              control = list(trace=1, parscale = c(1,0.1,0.1,1,1)))

}
\seealso{
\code{\link{auction_generate_data}}
}
